requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
pyyaml
ultralytics  # Para YOLOv8 y superiores